requirements.txt
cassandra-driver==3.29.2
click==8.1.7
Faker==33.0.0
geomet==0.2.1.post1
python-dateutil==2.9.0.post0
six==1.16.0
typing_extensions==4.12.2
